+++ lib/modules/text_editor/text_editor.dart
@@ -61,7 +61,7 @@ class TextEditorState extends State<TextEditor>
   late TextAlign align;
   late LayerBackgroundColorModeE backgroundColorMode;
   late double fontScale;
-  late TextStyle _customTextStyle;
+  late TextStyle selectedTextStyle;
   int _numLines = 0;
   double _colorPosition = 0;
 
@@ -72,7 +72,7 @@ class TextEditorState extends State<TextEditor>
     fontScale = textEditorConfigs.initFontScale;
     backgroundColorMode = textEditorConfigs.initialBackgroundColorMode;
 
-    _customTextStyle = widget.layer?.textStyle ??
+    selectedTextStyle = widget.layer?.textStyle ??
         textEditorConfigs.customTextStyles?.first ??
         const TextStyle();
     _initializeFromLayer();
@@ -245,6 +245,14 @@ class TextEditorState extends State<TextEditor>
     );
   }
 
+  /// Update the current text style.
+  void setTextStyle(TextStyle style) {
+    setState(() {
+      selectedTextStyle = style;
+      widget.onUpdateUI?.call();
+    });
+  }
+
   /// Closes the editor without applying changes.
   void close() {
     Navigator.pop(context);
@@ -262,7 +270,7 @@ class TextEditorState extends State<TextEditor>
           fontScale: fontScale,
           colorMode: backgroundColorMode,
           colorPickerPosition: _colorPosition,
-          textStyle: _customTextStyle,
+          textStyle: selectedTextStyle,
           // fontFamily: 'Roboto',
         ),
       );
@@ -410,6 +418,9 @@ class TextEditorState extends State<TextEditor>
 
   /// Builds the body of the text editor.
   Widget _buildBody() {
+    double barPickerPadding =
+        imageEditorTheme.editorMode == ThemeEditorMode.simple ? 10 : 60;
+
     return GestureDetector(
       behavior: HitTestBehavior.translucent,
       onTap: done,
@@ -420,9 +431,7 @@ class TextEditorState extends State<TextEditor>
             alignment: Alignment.topRight,
             child: Padding(
               padding: EdgeInsets.symmetric(
-                vertical: imageEditorTheme.editorMode == ThemeEditorMode.simple
-                    ? 10
-                    : 60,
+                vertical: barPickerPadding,
               ),
               child: BarColorPicker(
                 configs: widget.configs,
@@ -434,6 +443,7 @@ class TextEditorState extends State<TextEditor>
                       MediaQuery.of(context).viewInsets.bottom -
                       kToolbarHeight -
                       kBottomNavigationBarHeight -
+                      barPickerPadding * 2 -
                       MediaQuery.of(context).padding.top,
                 ),
                 onPositionChange: (value) {
@@ -454,15 +464,12 @@ class TextEditorState extends State<TextEditor>
               ),
             ),
           ),
-          TextEditorBottomBar(
-            configs: widget.configs,
-            selectedStyle: _customTextStyle,
-            onFontChange: (style) {
-              setState(() {
-                _customTextStyle = style;
-              });
-            },
-          ),
+          customWidgets.bottomBarTextEditor ??
+              TextEditorBottomBar(
+                configs: widget.configs,
+                selectedStyle: selectedTextStyle,
+                onFontChange: setTextStyle,
+              ),
           if (imageEditorTheme.editorMode == ThemeEditorMode.whatsapp)
             WhatsAppTextAppBar(
               configs: widget.configs,
@@ -502,7 +509,7 @@ class TextEditorState extends State<TextEditor>
                       _textCtrl.text,
                       backgroundColor: _getBackgroundColor,
                       textAlign: align,
-                      style: _customTextStyle.copyWith(
+                      style: selectedTextStyle.copyWith(
                         color: _getTextColor,
                         fontSize: _getTextFontSize,
                         fontWeight: FontWeight.w400,
@@ -531,13 +538,13 @@ class TextEditorState extends State<TextEditor>
                           hintText: _textCtrl.text.isEmpty
                               ? i18n.textEditor.inputHintText
                               : '',
-                          hintStyle: _customTextStyle.copyWith(
+                          hintStyle: selectedTextStyle.copyWith(
                             color: imageEditorTheme.textEditor.inputHintColor,
                             fontSize: _getTextFontSize,
                             fontWeight: FontWeight.w400,
                             height: 1.35,
                           )),
-                      style: _customTextStyle.copyWith(
+                      style: selectedTextStyle.copyWith(
                         color: Colors.transparent,
                         fontSize: _getTextFontSize,
                         fontWeight: FontWeight.w400,